move firebase setup to firebase.js
